+++ src/app/(protected)/layout.tsx
@@ -16,7 +16,7 @@ export default function ProtectedLayout({ children }: { children: ReactNode }) {
   }, []);
   useEffect(() => {
     if (ready && !token) {
-      router.replace(`/signin?redirect=${encodeURIComponent(pathname)}`);
+      router.replace(`/login?redirect=${encodeURIComponent(pathname)}`);
     }
   }, [ready, token, pathname, router]);
