+++ src/app/(public)/login/schemas/defaultLoginValues.ts
@@ -1,0 +1,4 @@
+export const defaultLoginValues = {
+  workEmail: '',
+  password: '',
+};
