+++ src/app/(public)/login/schemas/loginSchema.ts
@@ -1,8 +1,8 @@
 import { z } from 'zod';
 
-export const signinSchema = z.object({
+export const loginSchema = z.object({
   workEmail: z.string().email('Invalid email'),
   password: z.string().min(1, 'Password is required'),
 });
 
-export type SigninFormData = z.infer<typeof signinSchema>;
+export type LoginFormData = z.infer<typeof loginSchema>;
