--- src/app/(public)/signin/schemas/defaultSigninValues.ts
@@ -1,4 +1,0 @@
-export const defaultSignupValues = {
-  workEmail: '',
-  password: '',
-};
